Move closed day/days to openinghours namespace
