+++ src/Model/Serializer/ValueObject/Calendar/ClosedDayNormalizer.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace CultuurNet\UDB3\Model\Serializer\ValueObject\Calendar;
 
-use CultuurNet\UDB3\Model\ValueObject\Calendar\ClosedDay;
+use CultuurNet\UDB3\Model\ValueObject\Calendar\OpeningHours\ClosedDay;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
 
 final class ClosedDayNormalizer implements NormalizerInterface
