+++ src/Model/ValueObject/Calendar/CalendarWithClosedDays.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace CultuurNet\UDB3\Model\ValueObject\Calendar;
 
+use CultuurNet\UDB3\Model\ValueObject\Calendar\OpeningHours\ClosedDays;
+
 interface CalendarWithClosedDays extends Calendar
 {
     public function getClosedDays(): ClosedDays;
